+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to this project will be documented in this file.
 - BearSSL panic on ESP8266 in rare conditions (#22017)
 - Crash when calling TasmotaSerial destructor when initialized with incorrect arguments
 - LVGL Added OpenHASP icons to font `montserrat-28`
+- Matter fail to report Shutter status if no shutter is configured in Tasmota
 
 ### Removed